dom/workers/test/test_closeOnGC.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for DOM Worker Threads</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var count = 0;

  function testWorker(queryString) {
    ++count;
    var worker = new Worker("closeOnGC_worker.js?" + queryString);
    worker.onmessage = function(event) {
      is(event.data, "ready");
      worker = null;
    }

    var interval = setInterval(function() {
      var xhr = new XMLHttpRequest();
      xhr.open("GET", "closeOnGC_server.sjs?" + queryString, false);
      xhr.send();
      if (xhr.responseText != "closed") {
        SpecialPowers.gc();
        return;
      }
      clearInterval(interval);
      ok(true, "xhr correctly closed");
      if (--count == 0) {
        SimpleTest.finish();
      }
    }, 500);

    return worker;
  }

  testWorker("white");
  var worker = testWorker("gray");
  worker.onerror = function() {};
  worker.onerror.foo = worker;
  worker = null;

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
